Move all test cases to test directory
