--- cache/tests/BUILD
@@ -1,13 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_binary(
-    name = "test_cache",
-    srcs = [
-        "test_cache.c",
-        "test_cache.h",
-    ],
-    deps = [
-        "//cache",
-        "@unity",
-    ],
-)
