+++ tests/iota_api_mock.hh
@@ -1,7 +1,7 @@
 #include "cclient/iota_client_core_api.h"
 #include "cclient/iota_client_extended_api.h"
 #include "gmock/gmock.h"
-#include "serializer/test/test_serializer.h"
+#include "tests/test_serializer.h"
 
 /**
  * Interface class for iota APIs which all return RC_OK
