+++ tests/test_cache.c
