+++ tests/test_serializer.c
